% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_processing.R
\name{cut_and_label}
\alias{cut_and_label}
\title{Cut and label function}
\usage{
cut_and_label(
  x,
  breaks,
  right = TRUE,
  sep = " t/m ",
  extend_inf = TRUE,
  break_steps = 1
)
}
\arguments{
\item{x}{An integer vector to be cut and labeled.}

\item{breaks}{Vector of labels.}

\item{right}{Logical, should right hand side of breaks be included in labels?}

\item{sep}{Character, the separator used for labels between break points.}

\item{extend_inf}{Logical, should breaks be extended to -Inf and Inf?}

\item{break_steps}{numeric, how exact should the break steps be (e.g. 1, .1, .01)}
}
\value{
The rounded value of \code{x} with \code{digits} digits.
}
\description{
Cuts a vector into breaks and labels the levels.
}
\examples{
cut_and_label(c(1, 5, 10), breaks = c(2, 5), right = TRUE, sep = "-", extend_inf = TRUE)
cut_and_label(c(1, 5, 10), breaks = c(2, 10), right = FALSE, sep = " t/m ", extend_inf = FALSE)
cut_and_label(c(.1, .2, .4, .7, .9), breaks = c(.3, 0.7), break_steps = .1)

dplyr::starwars |>
dplyr::mutate(height_category =
cut_and_label(height, breaks= c(100, 200), right = FALSE, sep = "-", extend_inf = TRUE))
}
