% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARAutils-package.R
\docType{package}
\name{DARAutils-package}
\alias{DARAutils}
\alias{DARAutils-package}
\title{DARAutils: Useful generic functions that can be used in DARApipelines}
\description{
DARAutils is a R package that offers generic function that could be useful in DARA pipelines
}
\author{
\strong{Maintainer}: DARAteam \email{dara-team-list@rivm.nl}

}
\keyword{internal}
