% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{formatter_cli}
\alias{formatter_cli}
\title{Custom formatter that serves as input to DARA log_formatter}
\usage{
formatter_cli(
  ...,
  .logcall = sys.call(),
  .topcall = sys.call(-1),
  .topenv = parent.frame()
)
}
\arguments{
\item{...}{argument to msg_linesep}

\item{.logcall}{NULL or error}

\item{.topcall}{NULL or error}

\item{.topenv}{NULL or error}
}
\description{
Custom formatter that serves as input to DARA log_formatter
}
