% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rds_to_object.R
\name{import_rds_to_object}
\alias{import_rds_to_object}
\title{Import rds file(s) from a specified path}
\usage{
import_rds_to_object(
  path,
  date = today(),
  skip_modification_check = FALSE,
  pattern = ".",
  recursive = FALSE
)
}
\arguments{
\item{path}{Path where rds files are located.}

\item{date}{Date that rds files should contain in their file name.}

\item{skip_modification_check}{Skip using a date to find a file. Default: FALSE.}

\item{pattern}{Optional character string defining a pattern in the filename.}

\item{recursive}{Logical. Should the listing recurse into directories?}
}
\description{
Import rds files from a specified path and optionally a date.
By default, rds files from today are imported.
}
\examples{
\dontrun{
import_rds_to_object(path = "./example", date = as.Date("2000-01-01"))
import_rds_to_object(path = "./example", skip_modification_check = TRUE)
}

}
