% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_processing.R
\name{inheriting_merge}
\alias{inheriting_merge}
\title{Union of a set of named lists/vectors in which the last occurrence for each
entry is returned.}
\usage{
inheriting_merge(..., call = parent.frame())
}
\arguments{
\item{...}{Named lists/vectors.}

\item{call}{The calling function used for warning tracing. Default = parent.frame().}
}
\value{
An object with the same class as the first input containing the
inherited union of the inputs.
}
\description{
Takes lists or vectors and returns the named
union. If a named entry occurs more than once only the last entry is
returned. Unnamed entries are never returned. This function can be useful
when defining a set of default argument and a varying list of arguments
that need to overwrite those defaults.

The object class of the first entry is used for the return value.
}
\examples{
default_args <- list(a = 1, b = "something")
overwriting_args <- list(a = 2, c = "foo")
overwriting_args2 <- list(a = 3, b = "else")

inheriting_merge()
inheriting_merge(default_args)
inheriting_merge(default_args, overwriting_args)
inheriting_merge(default_args, overwriting_args, overwriting_args2)

}
