% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_processing.R
\name{ISOweek}
\alias{ISOweek}
\title{Converts a date to ISOweek format (yyyy-Wxx)}
\usage{
ISOweek(date)
}
\arguments{
\item{date}{date or datetime: the date (or vector of dates) that you want to convert to ISOweek.}
}
\value{
A character in the form yyyy-Wxx, e.g. "2022-W52" "2021-W52" "2020-W12".
}
\description{
This function returns the year and the week of the year of a given date
according to ISO 8601.
Mimics the
\href{https://www.rdocumentation.org/packages/ISOweek/versions/0.6-2/topics/ISOweek}{ISOweek}
function from the
\href{https://www.rdocumentation.org/packages/ISOweek}{ISOweek package}, but
uses lubridate functions internally.
}
\examples{
library(ggplot2)
economics |>
dplyr::mutate(ISOweek = date |> ISOweek())


}
