% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_processing.R
\name{join_date_sequence}
\alias{join_date_sequence}
\title{Executes join function 'join_f' to data frames x and y}
\usage{
join_date_sequence(x, y, by, date_x, date_y, join_f = left_join, ...)
}
\arguments{
\item{x}{Data.frame / tibble, data to be joined.}

\item{y}{Data.frame / tibble, population data.}

\item{by}{Character, extra column to join x and y (besides date columns), e.g. "Veiligheidsregio".
NULL if there is no other column.}

\item{date_x}{Date vector, variable in x.}

\item{date_y}{Date vector, variable in y (population). Becomes 'from' date.}

\item{join_f}{Function, one of: left_join, right_join, inner_join, full_join.}

\item{...}{Optional.}
}
\value{
Joint data.frame of \code{x} and \code{y}. Attention: column \code{date_y} is added.
}
\description{
Executes join function 'join_f' to data frames x and y, where y is the population data.
Y has a 'date_y' variable that specifies from which moment the population is valid
}
\examples{
x <- data.frame(
  dates = seq.Date(
    from = as.Date("2020-01-01"),
    to = as.Date("2022-07-01"),
    by = "6 months"
  )
)
y <- tibble::tribble(
  ~Date_poll, ~Population, ~Country,
  as.Date("2020-01-01"), 17407585, "Netherlands",
  as.Date("2021-01-01"), 17475415, "Netherlands",
  as.Date("2022-01-01"), 17590672, "Netherlands"
)
join_date_sequence(x = x, y = y, by = NULL, date_x = dates, date_y = Date_poll)

}
