% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{layout_generator_subject}
\alias{layout_generator_subject}
\title{Adds date, time and subject to logger message}
\usage{
layout_generator_subject(subject)
}
\arguments{
\item{subject}{string, subject of logger}
}
\description{
Adds date, time and subject to logger message
}
