% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{local_logger_sink}
\alias{local_logger_sink}
\title{Turn off logging for the duration of the parent function's execution and restore}
\usage{
local_logger_sink()
}
\description{
This function is used in many test functions of DARA packages, because logging
can clutter the output of test runs in the Build tab.
}
