% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{send_mail}
\alias{send_mail}
\title{Send mail}
\usage{
send_mail(
  subject,
  body,
  recipients,
  is.html = FALSE,
  sender = "noreply@domain.nl",
  attachments = NULL,
  cc = NULL,
  bcc = NULL
)
}
\arguments{
\item{subject}{Character. Subject of email.}

\item{body}{Character. Body of email.}

\item{recipients}{List of characters. Mail addresses of recipients.}

\item{is.html}{Boolean. Should content type of email be html?}

\item{sender}{Character. Sender of email.}

\item{attachments}{Vector of characters. Paths to attachments.}

\item{cc}{Character. CC mail addresses.}

\item{bcc}{Character. BCC mail addresses.}
}
\description{
Sends an automated e-mail. Can be useful after running your pipeline,
e.g. succes or failure.
}
\examples{
\dontrun{
send_mail(
  subject = "Test mail",
  body = "This is a test e-mail",
  recipients = "hello[@]example.com"
)
}

}
