% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade_config}
\alias{upgrade_config}
\title{Converts the old-style config file to the new multiple config files}
\usage{
upgrade_config(
  config_file_old = "config/config.yaml",
  config_dir_new = "config"
)
}
\arguments{
\item{config_file_old}{String. Location of the old config file.}

\item{config_dir_new}{String. Location of the new config folder.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function to upgrade from the singular old style config file to
the new multi-file config files. The new config files enables the use
functions such as pipeline_init(),  pipeline_run() and
pipeline_import_for().
}
\examples{
\dontrun{
upgrade_config()
}

}
