% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_processing.R
\name{xlabel_weekyear}
\alias{xlabel_weekyear}
\title{xlabel_weekyear}
\usage{
xlabel_weekyear(isoweeks)
}
\arguments{
\item{isoweeks}{character vector with ISOweeks.}
}
\description{
Given a character vector of ISOweeks, labels first ISOweek of a year with the
year-number and all others without the year-number, thus only with the week-
number.
}
\examples{
isoweeks <- ISOweek(seq.Date(from = as.Date("2023-01-01"), to = as.Date("2023-02-01"), by = "week"))
xlabel_weekyear(isoweeks = isoweeks)

}
